requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
Pillow>=8.3.1
opencv-python>=4.5.3
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.2
scikit-learn>=0.24.2
python-Levenshtein>=0.12.2
timm>=0.9.0